+++ crates/on-chain/src/lib.rs
@@ -2,5 +2,5 @@
 // #![warn(missing_docs)]
 #![warn(unsafe_code)]
 
-pub mod architect;
-pub mod clairvoyance;
+pub mod executor;
+pub mod monitor;
